tibber-case/pyproject.toml
[tool.poetry]
name = "tibber-case"
version = "0.1.0"
description = "Tibber Data Engineering Task"
authors = ["ramtin95 <ramme95@gmail.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11.9"
pandas = "^2.2.3"
requests = "^2.32.3"
sqlalchemy = "^2.0.39"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "0.9.9"
mypy = "^1.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

lint.extend-select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "B",   # flake8-bugbear
    "I",   # isort
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
]

# Files to exclude
exclude = [
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
    "*.sql",        # Exclude all SQL files
    "**/*.sql",     # Exclude SQL files in subdirectories
]

# Assume Python 3.11
target-version = "py311"

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true